[csv] move spirit usage to cpp
