+++ plugins/input/csv/build.py
@@ -39,6 +39,7 @@ else:
 
     plugin_sources = Split(
         """
+        %(PLUGIN_NAME)s_utils.cpp
         %(PLUGIN_NAME)s_datasource.cpp
         %(PLUGIN_NAME)s_featureset.cpp
         %(PLUGIN_NAME)s_inline_featureset.cpp
