+++ plugins/input/csv/csv_datasource.cpp
@@ -37,9 +37,12 @@
 #include <mapnik/memory_featureset.hpp>
 #include <mapnik/boolean.hpp>
 #include <mapnik/util/trim.hpp>
+#include <mapnik/geometry.hpp>
+#include <mapnik/geometry_adapters.hpp>
 #include <mapnik/util/geometry_to_ds_type.hpp>
 #include <mapnik/value_types.hpp>
 #include <mapnik/util/fs.hpp>
+#include <mapnik/make_unique.hpp>
 #include <mapnik/util/spatial_index.hpp>
 #include <mapnik/geom_util.hpp>
 #if defined(MAPNIK_MEMORY_MAPPED_FILE)
